Add test for fixer
With `RunWithSuggestedFixes` we can assert that the output of the
diagnostics created are corect. It's a bit weird because the golden file
will contain all the expected assertions even though the would no longer
be triggered. For now they're kept to make the test pass.

Ref: https://godocs.io/golang.org/x/tools/go/analysis/analysistest#RunWithSuggestedFixes
